+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Other
 
 * Removes unused type `ConflictingEvidenceError` (#87)
+* Adds profiling for variable elimination (#89)
 
 ## [[v0.6.0]](https://github.com/mlange-42/bbn/compare/v0.5.0...v0.6.0)